Add multi-language support to the portfolio website
The portfolio website has been updated to support multiple languages, specifically English and Japanese. The language context was defined using the React context API. After that, sections which require translation (About, Education, and Projects) were refactored to conditionally render content based on the current language state. In addition, a language Switch component has been created and implemented as a floating button on the layout, giving users the ability to switch between languages.

This update makes the portfolio website more accessible to Japanese and allows users to view the content in their preferred language.